SQLQueries.sql: SQL queries for analyzing COVID-19 data. exploring total cases, total deaths, death percentages, population vs. vaccination rates, and more.
This SQL file contains a set of queries to analyze COVID-19 data from the my local PortfolioProject1 database. These queries cover various aspects of the pandemic, such as total cases, total deaths, death percentages, population vs. vaccination rates, and more. The queries are designed to provide insights into the impact of COVID-19 at the global, regional, and local levels.

The file begins with queries that retrieve data from the CovidDeaths and CovidVaccinations tables, filtering out rows with null continent values. The queries include sorting and filtering criteria to focus on specific regions and explore interesting patterns and trends.

The file also showcases the use of common table expressions (CTEs) to calculate rolling sums of new vaccinations over time, as well as the percentage of the population vaccinated. Furthermore, it demonstrates the creation of temporary tables and views for further analysis and visualization purposes.

Please note that these queries assume the existence of the local PortfolioProject1 database and the specified table names. Adjustments may be required to fit your specific database structure and naming conventions.